+++ frontend/src/App.tsx
@@ -1,0 +1,11 @@
+import './App.css'
+
+function App() {
+    return (
+        <div id="App">
+            <h1>Athena</h1>
+        </div>
+    )
+}
+
+export default App